Separate common and analysis-specific selections
