+++ preselection/src/corrections.cpp
@@ -1,5 +1,43 @@
 #include "corrections.h"
 
+// BTAGGING WORKING POINTS (TO BE UPDATED)
+// float looseDFBtagWP(std::string year){
+//     if(year == "2016preVFP")
+//         return 0.0508;
+//     if(year == "2016postVFP")
+//         return 0.0480;
+//     if(year == "2017")
+//         return 0.0532;
+//     if(year == "2018")
+//         return 0.0490;
+//     return -1;
+// }
+
+// float mediumDFBtagWP(std::string year){
+//     if(year == "2016preVFP")
+//         return 0.2598;
+//     if(year == "2016postVFP")
+//         return 0.2489;
+//     if(year == "2017")
+//         return 0.3040;
+//     if(year == "2018")
+//         return 0.2783;
+//     return -1;
+// }
+
+// float tightDFBtagWP(std::string year){
+//     if(year == "2016preVFP")
+//         return 0.6502;
+//     if(year == "2016postVFP")
+//         return 0.6377;
+//     if(year == "2017")
+//         return 0.7476;
+//     if(year == "2018")
+//         return 0.7100;
+//     return -1;
+// }
+
+
 // CORRECTIONS
 RNode defineCorrectedCols(RNode df) {
     return df.Define("CorrJet_pt", "Jet_pt")
@@ -9,6 +47,7 @@ RNode defineCorrectedCols(RNode df) {
             .Define("CorrMET_pt", "MET_pt")
             .Define("CorrMET_phi", "MET_phi");
 }
+
 
 // RNode JMS_Corrections(correction::CorrectionSet cset_jet_mass_scale, RNode df, std::string variation) { 
 //     auto eval_correction = [cset_jet_mass_scale, variation] (std::string year, float mass) {
